+++ core/src/main/kotlin/fr/xgouchet/elmyr/annotation/RegexForgery.kt
@@ -3,6 +3,9 @@ package fr.xgouchet.elmyr.annotation
 /**
  * Mark a field, property or method parameter as a Regex based String forgery.
  * @param value the regex pattern to match.
+ *
+ * @deprecated Regex based string forgery can now be created using the StringForgery annotation
+ * @see [StringForgery]
  */
 @Target(
         AnnotationTarget.VALUE_PARAMETER,
@@ -10,6 +13,7 @@ package fr.xgouchet.elmyr.annotation
         AnnotationTarget.PROPERTY
 )
 @Retention(AnnotationRetention.RUNTIME)
+@Deprecated("Regex based string forgery can now be created using the StringForgery annotation")
 @MustBeDocumented
 annotation class RegexForgery(
     val value: String