+++ core/src/main/kotlin/fr/xgouchet/elmyr/annotation/StringForgery.kt
@@ -5,10 +5,11 @@ import fr.xgouchet.elmyr.Case
 /**
  * Mark a field, property or method parameter as a String forgery.
  *
- * @param value the [StringForgeryType] of String to forge ([StringForgeryType.ALPHABETICAL] by default)
+ * @param type the [StringForgeryType] of String to forge ([StringForgeryType.ALPHABETICAL] by default)
  * @param case the case to use ([Case.ANY] by default). This will only be used for the following types :
  * [StringForgeryType.ALPHABETICAL], [StringForgeryType.ALPHA_NUMERICAL], [StringForgeryType.NUMERICAL],
  * [StringForgeryType.HEXADECIMAL]
+ * @param regex the regex pattern to match (leave empty to use the type and case instead).
  */
 @Target(
         AnnotationTarget.VALUE_PARAMETER,
@@ -18,6 +19,7 @@ import fr.xgouchet.elmyr.Case
 @Retention(AnnotationRetention.RUNTIME)
 @MustBeDocumented
 annotation class StringForgery(
-    val value: StringForgeryType = StringForgeryType.ALPHABETICAL,
-    val case: Case = Case.ANY
+    val type: StringForgeryType = StringForgeryType.ALPHABETICAL,
+    val case: Case = Case.ANY,
+    val regex: String = ""
 )